Only install from  mypyc-requirements.txt if it exists
This will let us handle the transition to a merged repo

+++ config.sh
@@ -1,6 +1,8 @@
 function pre_build {
     pip install virtualenv
-    pip install -r mypy/mypyc-requirements.txt
+    if [ -f mypy/mypyc-requirements.txt ]; then
+	pip install -r mypy/mypyc-requirements.txt
+    fi
 }
 
 function run_tests {
